[EI-300] [web] refactor: Update MetadataList import in metadata-list.tsx
