+++ web/app/ui/components/numerical-guidance/indicator-board-metadata/metadata-list-item/indicator-unit-selector.tsx
