+++ web/app/ui/components/numerical-guidance/indicator-board-metadata/metadata-list-item/metadata-list-item-row.tsx
